backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0
praw==7.7.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
werkzeug==2.2.3
sqlalchemy==2.0.25
openai==1.11.0 